Restore Web UI external access feature
